refactor: move the auth and user features out from under src/modules to src/
