+++ packages/api/@types/express/index.d.ts
@@ -1,4 +1,4 @@
-import User from '../../src/modules/user/user.entity';
+import User from '../../src/user/user.entity';
 
 declare global {
     namespace Express {
