+++ packages/api/src/common/auth.middleware.ts
@@ -1,8 +1,8 @@
 import { Injectable, NestMiddleware } from '@nestjs/common';
 import { Request, Response, NextFunction } from 'express';
 import clc from 'cli-color';
-import { AuthService } from '../modules/auth/auth.service';
-import { UserService } from '../modules/user/user.service';
+import { AuthService } from '../auth/auth.service';
+import { UserService } from '../user/user.service';
 
 @Injectable()
 export class AuthMiddleware implements NestMiddleware {
