+++ packages/api/test/auth.e2e-spec.ts
@@ -4,9 +4,9 @@ import { INestApplication } from '@nestjs/common';
 import { TestingModule, Test } from '@nestjs/testing';
 import request from 'supertest';
 import { AppModule } from '../src/app.module';
-import { AuthService } from '../src/modules/auth/auth.service';
-import { CreateUserDto } from '../src/modules/user/create-user.dto';
-import { UserService } from '../src/modules/user/user.service';
+import { AuthService } from '../src/auth/auth.service';
+import { CreateUserDto } from '../src/user/create-user.dto';
+import { UserService } from '../src/user/user.service';
 
 const oliver: CreateUserDto = {
     firstName: 'Ollie',
